Refactored ChangeDueDate fixtures into new fixture Class
